Adapt run metrics panel to the new dashboard
