+++ sematic/ui/packages/common/src/sockets.ts
@@ -1,3 +1,5 @@
 import io from "socket.io-client";
 
 export const graphSocket = io("/graph");
+export const pipelineSocket = io("/pipeline");
+export const metricsSocket = io("/metrics");
